Fix schema update (#18253)
* Add Filter and change to default entitymanagers

* Add service to config

* Add fix for last insertid

* fix line length

* fix gh runner

* Use ubuntu-latest for static analysis workflow

Updated the GitHub Actions workflow to use the latest Ubuntu version instead of a specific version. This ensures the workflow benefits from up-to-date tools and security patches.

* add doc, fix composer requirements

* Update return type in docblock for upsert method

The return type in the upsert method docblock was changed to include `int|string|null` for better accuracy and clarity. This ensures the documentation reflects the actual return values of the method.

* fix getTableName

---------

Co-authored-by: Herbert Roth <herbert.roth@pimcore.com>