Improvement: check validity for (Multi)Select before setting values (#17742)
* Check validity for (Multi)Select before setting values

When using the Pimcore API, any value can be set on a select or
multiselect field. This improvement checks the validity before setting
the data.

It applies to configured options, SelectOptions and OptionProviders.

Resolves: #17343

* Apply php-cs-fixer changes

* Skip the Select validation if no options are defined yet

In case no list of options is defined (yet), don't validate the new
input value. This avoids large volume of ValidationErrors when for
example a SelectOption or OptionsProvider service gets missing

* Update the codeception test for select data types

- use actual values instead of numbers for the MultiSelect fields (as
  the Country and Language (single) Select already do)
- fix the faulty assert for LanguageMultiSelect (which was using the
  CountryMultiSelect assert)

* Apply php-cs-fixer changes

* Add doc block to (Multi)Select getOption() methods

As the return type is nullable, this should be indicated by the doc
block as well.

* Apply php-cs-fixer changes

* Remove fallback to empty array on Select validity check

As the options were already checked for being empty, the array will
never need a fallback at this point

* Avoid using empty() on nullable arrays in validitycheck on Selects

Co-authored-by: Sebastian Blank <sebastian.bl@gmx.de>

* Change isValidOption() scope to private

Co-authored-by: Sebastian Blank <sebastian.bl@gmx.de>

* add enforce validation setting

* add test

* test for default behavior

* Update README.md

* fix

* fix typo

* fix for localizedfield

---------

Co-authored-by: samynw <samynw@users.noreply.github.com>
Co-authored-by: Sebastian Blank <sebastian.bl@gmx.de>
Co-authored-by: Ji Jia Jia <kingjia90@gmail.com>